FabriX/requirements.txt
streamlit==1.35.0
streamlit-drawable-canvas==0.9.4
Pillow==10.3.0
numpy==1.26.4
openai==1.30.1